Fixed/Added AI for: Focus Punch, Roar, Future Sight, Assist, (fixed U-Turn percentages... hopefully), Quash, Pursuit, Pain Split, Two Turn Attacks, Salt Cure. Fixed + added tests based on some of the above

+++ include/test/battle.h
@@ -700,7 +700,7 @@ struct ExpectedAiScore
     u8 moveSlot1:2;
     u8 moveSlot2:2;
     u8 target:2;
-    s8 value; // value
+    s16 value; // value
     u8 cmp:3; // Uses battle script command's CMP_ macros
     u8 toValue:1; // compare to value, not to move
     u8 set:1;
